Add eslint (air-bnb)
