+++ src/config/i18n/i18n.js
@@ -5,17 +5,16 @@ import Backend from 'i18next-http-backend';
 import LanguageDetector from 'i18next-browser-languagedetector';
 
 i18n
-    .use(Backend)
-    .use(LanguageDetector)
-    .use(initReactI18next)
-    .init({
-        fallbackLng: 'ru',
-        debug: false, // Дебаг сообщения в консоль
+  .use(Backend)
+  .use(LanguageDetector)
+  .use(initReactI18next)
+  .init({
+    fallbackLng: 'ru',
+    debug: false, // Дебаг сообщения в консоль
 
-        interpolation: {
-            escapeValue: false, // not needed for react as it escapes by default
-        }
-    });
-
+    interpolation: {
+      escapeValue: false, // not needed for react as it escapes by default
+    },
+  });
 
 export default i18n;
