--- src/index.js
@@ -1,15 +1,0 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import { BrowserRouter } from 'react-router-dom'
-import App from './App/App'
-import './index.css'
-import i18n from "./config/i18n/i18n";
-
-const root = ReactDOM.createRoot(document.getElementById('root'))
-root.render(
-  <BrowserRouter>
-    <React.StrictMode>
-      <App />
-    </React.StrictMode>
-  </BrowserRouter>
-)
